requirements.txt
boto==2.38.0
Collectfast==0.2.3
conekta==1.2
Django==1.8.4
django-cors-headers==1.1.0
django-countries==3.3
django-environ==0.3.0
django-filter==0.11.0
django-modeltranslation==0.10.1
django-solo==1.1.0
django-storages-redux==1.3
django-taggit==0.17.1
django-taggit-serializer==0.1.5
django-uuslug==1.1.6
djangorestframework==3.2.3
httplib2==0.9.1
Markdown==2.6.2
nose==1.3.7
Pillow==2.9.0
psycopg2==2.6.1
python-slugify==1.1.4
pytz==2015.4
simplejson==3.8.0
Unidecode==0.4.18
